Added dark and light mode with toggles
Added an light and dark mode with a toggle to adjust for different situations (dark mode looks so cool) as well as adjusting the efficiency of the tool box. Next stage is to get the data to sync to the database information on heating types and add another toolbox for toggles.